[refactor] Move Rating model from accounts app to movies app
