+++ backend/accounts/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
-from .models import User, Rating
+from .models import User
 from django.contrib.auth.admin import UserAdmin
 
 
 admin.site.register(User, UserAdmin)
-admin.site.register(Rating)
